:white_check_mark:  fix tests (circular dependency)
