+++ src/modules/candidateNotification/index.ts
@@ -1,4 +1,5 @@
 export * from './eventHandlers';
 export * from './events';
+export * from './helpers';
 export * from './queries';
 export * from './CandidateNotification';
