+++ src/modules/project/events/ProjectCertificateGenerated.ts
@@ -3,7 +3,7 @@ import {
   BaseDomainEventProps,
   DomainEvent,
 } from '../../../core/domain/DomainEvent'
-import { makeCandidateNotificationId } from '../../candidateNotification/CandidateNotification'
+import { makeCandidateNotificationId } from '../../candidateNotification/helpers'
 
 export interface ProjectCertificateGeneratedPayload {
   certificateFileId: string
