+++ src/modules/projectClaim/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './makeClaimProjectAggregateId';
