+++ src/modules/projectClaim/useCases/claimProject.ts
@@ -3,7 +3,8 @@ import { User } from '../../../entities'
 import { AggregateHasBeenUpdatedSinceError } from '../../shared'
 import { EventBus, Repository, TransactionalRepository, UniqueEntityID } from '../../../core/domain'
 import { FileContents, FileObject, IllegalFileDataError, makeFileObject } from '../../file'
-import { makeClaimProjectAggregateId, ProjectClaim } from '../ProjectClaim'
+import { makeClaimProjectAggregateId } from '../helpers'
+import { ProjectClaim } from '../ProjectClaim'
 import { GetProjectDataForProjectClaim } from '../../project/queries/GetProjectDataForProjectClaim'
 import { ProjectClaimFailed } from '../events'
 import { ClaimerIdentityCheckHasFailedError } from '..'
